add `UTXOType` field and parsing utilities
Add UTXOType field for determining if a UTXO is eligible for shared
secret derivation.

Add tests/utils.ts for parsing BIP352 test vectors. These utils
are used to determine the UTXOType of the input.

The utilities are only added for the tests and not the general library since
in normal usage the sending wallet already knows the UTXOType and can
set it directly.